requirements.txt
Flask==3.0.0
python-amazon-paapi==5.0.1
python-dotenv==1.0.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
pytest==7.4.3
gunicorn==21.2.0
